Revert "remove definitions from abstract class"
This reverts commit 420fc63c20e8d72b83fe63fb7481c29b511daf07.
