+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/SchemaDefinitionsAltair.java
@@ -34,8 +34,6 @@ import tech.pegasys.teku.spec.datastructures.lightclient.LightClientUpdateSchema
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.metadata.versions.altair.MetadataMessageSchemaAltair;
 import tech.pegasys.teku.spec.datastructures.operations.AggregateAndProof.AggregateAndProofSchema;
 import tech.pegasys.teku.spec.datastructures.operations.AttestationSchema;
-import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashingSchema;
-import tech.pegasys.teku.spec.datastructures.operations.IndexedAttestationSchema;
 import tech.pegasys.teku.spec.datastructures.operations.SignedAggregateAndProof.SignedAggregateAndProofSchema;
 import tech.pegasys.teku.spec.datastructures.operations.versions.altair.ContributionAndProofSchema;
 import tech.pegasys.teku.spec.datastructures.operations.versions.altair.SignedContributionAndProofSchema;
@@ -43,8 +41,6 @@ import tech.pegasys.teku.spec.datastructures.operations.versions.altair.SyncAggr
 import tech.pegasys.teku.spec.datastructures.operations.versions.altair.SyncCommitteeContributionSchema;
 import tech.pegasys.teku.spec.datastructures.operations.versions.altair.SyncCommitteeMessageSchema;
 import tech.pegasys.teku.spec.datastructures.operations.versions.phase0.AttestationPhase0Schema;
-import tech.pegasys.teku.spec.datastructures.operations.versions.phase0.AttesterSlashingPhase0Schema;
-import tech.pegasys.teku.spec.datastructures.operations.versions.phase0.IndexedAttestationPhase0Schema;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateSchema;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.altair.BeaconStateAltair;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.altair.BeaconStateSchemaAltair;
@@ -52,8 +48,6 @@ import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.altair.M
 
 public class SchemaDefinitionsAltair extends AbstractSchemaDefinitions {
   private final AttestationSchema<?> attestationSchema;
-  private final IndexedAttestationPhase0Schema indexedAttestationSchema;
-  private final AttesterSlashingPhase0Schema attesterSlashingSchema;
   private final SignedAggregateAndProofSchema signedAggregateAndProofSchema;
   private final AggregateAndProofSchema aggregateAndProofSchema;
   private final BeaconStateSchemaAltair beaconStateSchema;
@@ -71,12 +65,7 @@ public class SchemaDefinitionsAltair extends AbstractSchemaDefinitions {
 
   public SchemaDefinitionsAltair(final SpecConfigAltair specConfig) {
     super(specConfig);
-    final long maxValidatorsPerAttestation = getMaxValidatorPerAttestation(specConfig);
-    this.indexedAttestationSchema = new IndexedAttestationPhase0Schema(maxValidatorsPerAttestation);
-    this.attesterSlashingSchema =
-        new AttesterSlashingPhase0Schema(
-            indexedAttestationSchema.castTypeToIndexedAttestationSchema());
-    this.attestationSchema = new AttestationPhase0Schema(maxValidatorsPerAttestation);
+    this.attestationSchema = new AttestationPhase0Schema(getMaxValidatorPerAttestation(specConfig));
     this.aggregateAndProofSchema = new AggregateAndProofSchema(attestationSchema);
     this.signedAggregateAndProofSchema = new SignedAggregateAndProofSchema(aggregateAndProofSchema);
     this.beaconStateSchema = BeaconStateSchemaAltair.create(specConfig);
@@ -84,7 +73,7 @@ public class SchemaDefinitionsAltair extends AbstractSchemaDefinitions {
         BeaconBlockBodySchemaAltairImpl.create(
             specConfig,
             getAttesterSlashingSchema(),
-            maxValidatorsPerAttestation,
+            getMaxValidatorPerAttestation(specConfig),
             "BeaconBlockBodyAltair");
     this.beaconBlockSchema = new BeaconBlockSchema(beaconBlockBodySchema, "BeaconBlockAltair");
     this.signedBeaconBlockSchema =
@@ -184,16 +173,6 @@ public class SchemaDefinitionsAltair extends AbstractSchemaDefinitions {
   @Override
   public BeaconBlockBodyBuilder createBeaconBlockBodyBuilder() {
     return new BeaconBlockBodyBuilderAltair(beaconBlockBodySchema);
-  }
-
-  @Override
-  public IndexedAttestationSchema<?> getIndexedAttestationSchema() {
-    return indexedAttestationSchema;
-  }
-
-  @Override
-  public AttesterSlashingSchema<?> getAttesterSlashingSchema() {
-    return attesterSlashingSchema;
   }
 
   @Override
